Algunos cambios en Entidades para que matcheen bien con el tipo de dato de SQL. Listado de Afiliados (revisar esto, algunos valores importantes me vienen en null). Ver de modificar el SP de Afiliados para que haga un JOIN con PlanesMedicos y me devuelva el nombre del Plan

+++ ClinicaFrba-BETTER_CALL_JUAN/ClinicaFrba/Modelo/Objetos/BajaPaciente.cs
@@ -8,8 +8,8 @@ namespace ClinicaFrba
 {
     class BajaPaciente
     {
-        public int id { get; set; }
-        public int pacienteId { get; set; }
+        public decimal id { get; set; }
+        public decimal pacienteId { get; set; }
         public DateTime fechaBaja { get; set; }
     }
 }
