+++ ClinicaFrba-BETTER_CALL_JUAN/ClinicaFrba/Modelo/Objetos/Medico.cs
@@ -8,17 +8,17 @@ namespace ClinicaFrba
 {
     class Medico
     {
-        public int matricula { get; set; }
+        public decimal matricula { get; set; }
         public string nombre { get; set; }
         public string apellido { get; set; }
         public string tipoDoc { get; set; }
         public int nroDoc { get; set; }
         public string direccion { get; set; }
-        public int telefono { get; set; }
+        public decimal telefono { get; set; }
         public string mail { get; set; }
         public DateTime fechaNacimiento { get; set; }
         public char sexo { get; set; }
-        public int usuarioId { get; set; }
+        public decimal usuarioId { get; set; }
 
     }
 }
