+++ ClinicaFrba-BETTER_CALL_JUAN/ClinicaFrba/Modelo/Objetos/TipoEspecialidad.cs
@@ -8,7 +8,7 @@ namespace ClinicaFrba
 {
     class TipoEspecialidad
     {
-        public int codigo { get; set; }
+        public decimal codigo { get; set; }
         public string descripcion { get; set; }
     }
 }
